requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
SQLAlchemy>=2.0.36
Werkzeug>=3.1.3
openai>=1.58.1
PyMuPDF>=1.25.1
Pillow>=11.1.0
python-dotenv>=1.0.1
gunicorn>=23.0.0
# psycopg2-binary is only needed for PostgreSQL (production/Render)
# Uncomment below when deploying with PostgreSQL:
# psycopg2-binary>=2.9.10
